Create FUNDING.yml
As we progress through our development stages, we would be honored and grateful for any and all gifts of funding that help us complete the demo version of "Tiny Monsters: Jungle Mummies" as well as any of our other demo projects currently in development. We are striving to create unique and exciting entertainment for gamers. We're pretty sure we've got it down... 

Needless to say, any sponsorship or funding will not only be greatly appreciated but will ensure a free copy of the final published game as well as special (sponsor only) free DLC.